+++ src/Umbraco.Web.UI.Client/src/packages/media/media/search/index.ts
@@ -1,1 +1,2 @@
 export * from './media.search-provider.js';
+export type * from './types.js';
