+++ src/Umbraco.Web.UI.Client/src/packages/media/media/search/media-search.repository.ts
@@ -1,9 +1,9 @@
 import { UmbMediaSearchServerDataSource } from './media-search.server.data-source.js';
-import type { UmbMediaSearchItemModel } from './media.search-provider.js';
 import type { UmbSearchRepository, UmbSearchRequestArgs } from '@umbraco-cms/backoffice/search';
 import { UmbControllerBase } from '@umbraco-cms/backoffice/class-api';
 import type { UmbControllerHost } from '@umbraco-cms/backoffice/controller-api';
 import type { UmbApi } from '@umbraco-cms/backoffice/extension-api';
+import type { UmbMediaSearchItemModel } from './types.js';
 
 export class UmbMediaSearchRepository
 	extends UmbControllerBase
