+++ src/Umbraco.Web.UI.Client/src/packages/media/media/search/media-search.server.data-source.ts
@@ -1,9 +1,9 @@
 import { UMB_MEDIA_ENTITY_TYPE } from '../entity.js';
-import type { UmbMediaSearchItemModel } from './media.search-provider.js';
 import type { UmbSearchDataSource, UmbSearchRequestArgs } from '@umbraco-cms/backoffice/search';
 import type { UmbControllerHost } from '@umbraco-cms/backoffice/controller-api';
 import { MediaService } from '@umbraco-cms/backoffice/external/backend-api';
 import { tryExecuteAndNotify } from '@umbraco-cms/backoffice/resources';
+import type { UmbMediaSearchItemModel } from './types.js';
 
 /**
  * A data source for the Rollback that fetches data from the server
