+++ src/Umbraco.Web.UI.Client/src/packages/media/media/search/media.search-provider.ts
@@ -3,10 +3,6 @@ import { UmbMediaSearchRepository } from './media-search.repository.js';
 import type { UmbSearchProvider, UmbSearchRequestArgs } from '@umbraco-cms/backoffice/search';
 import { UmbControllerBase } from '@umbraco-cms/backoffice/class-api';
 
-export interface UmbMediaSearchItemModel extends UmbMediaItemModel {
-	href: string;
-}
-
 export class UmbMediaSearchProvider extends UmbControllerBase implements UmbSearchProvider<UmbMediaSearchItemModel> {
 	#repository = new UmbMediaSearchRepository(this);
