+++ src/Umbraco.Web.UI.Client/src/packages/media/media/search/types.ts
@@ -1,0 +1,5 @@
+import type { UmbMediaItemModel } from '../repository/index.js';
+
+export interface UmbMediaSearchItemModel extends UmbMediaItemModel {
+	href: string;
+}
